+++ discretisedfield/field_rotator.py
@@ -167,7 +167,8 @@ class FieldRotator:
 
         self._rotated_field = df.Field(mesh=new_mesh,
                                        dim=self._orig_field.dim,
-                                       value=new_m)
+                                       value=new_m,
+                                       dtype=self._orig_field.dtype)
 
     def clear_rotation(self):
         """Remove all rotations."""
